+++ ForeCastApp/app/src/main/java/com/example/forecastmvvm/data/db/unitlocalized/ImperialCurrentWeatherEntry.kt
@@ -4,23 +4,23 @@ import androidx.room.ColumnInfo
 
 class ImperialCurrentWeatherEntry(
     @ColumnInfo(name = "feelslike")
-    override val feelslike: Int,
+    override val feelslike: Double,
     @ColumnInfo(name = "isDay")
     override val isDay: String,
     @ColumnInfo(name = "precip")
-    override val precip: Int,
+    override val precip: Double,
     @ColumnInfo(name = "temperature")
-    override val temperature: Int,
+    override val temperature: Double,
     @ColumnInfo(name = "visibility")
-    override val visibility: Int,
+    override val visibility: Double,
     @ColumnInfo(name = "weatherDescriptions")
     override val weatherDescriptions: List<String>,
     @ColumnInfo(name = "weatherIcons")
     override val weatherIcons: List<String>,
     @ColumnInfo(name = "windDegree")
-    override val windDegree: Int,
+    override val windDegree: Double,
     @ColumnInfo(name = "windDir")
     override val windDir: String,
     @ColumnInfo(name = "windSpeed")
-    override val windSpeed: Int
+    override val windSpeed: Double
 ) : UnitSpecificCurrentWeatherEntry
