+++ ForeCastApp/app/src/main/java/com/example/forecastmvvm/data/db/unitlocalized/UnitSpecificCurrentWeatherEntry.kt
@@ -3,14 +3,14 @@ package com.example.forecastmvvm.data.db.unitlocalized
 import com.google.gson.annotations.SerializedName
 
 interface UnitSpecificCurrentWeatherEntry {
-    val feelslike: Int
+    val feelslike: Double
     val isDay: String
-    val precip: Int
-    val temperature: Int
-    val visibility: Int
+    val precip: Double
+    val temperature: Double
+    val visibility: Double
     val weatherDescriptions: List<String>
     val weatherIcons: List<String>
-    val windDegree: Int
+    val windDegree: Double
     val windDir: String
-    val windSpeed: Int
+    val windSpeed: Double
 }
